devop: refactor token info module
